requirements.txt
appnope==0.1.3
apyori==1.1.2
asttokens==2.4.1
certifi==2024.6.2
cffi==1.15.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
fonttools==4.49.0
idna==3.7
ipykernel==6.29.0
ipython==8.21.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mip==1.14.2
mlxtend==0.23.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
PySCIPOpt==5.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.32.3
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
tzdata==2023.3
urllib3==2.2.2
wcwidth==0.2.13
